pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mollie-cli"
description = 'Simple commandline client for Mollie'
readme = "README.md"
requires-python = ">=3.7"
license = "MIT"
keywords = []
authors = [
  { name = "Tom Hendrikx", email = "tom@whyscream.net" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "click>=8.1.3,<9",
  "mollie-api-python>=3.0.0rc1,<4",
  "tabulate>=0.9.0",
]
dynamic = ["version"]

[project.scripts]
mollie-cli = "mollie_cli.cli:main"

[project.urls]
Documentation = "https://github.com/whyscream/mollie-cli#readme"
Issues = "https://github.com/whyscream/mollie-cli/issues"
Source = "https://github.com/whyscream/mollie-cli"

[tool.hatch.version]
path = "mollie_cli/__about__.py"

[tool.hatch.envs.default]
dependencies = [
  "pytest",
  "pytest-cov",
]
[tool.hatch.envs.default.scripts]
cov = "pytest --cov-report=term-missing --cov-config=pyproject.toml --cov=mollie_cli --cov=tests {args}"
no-cov = "cov --no-cov {args}"

[[tool.hatch.envs.test.matrix]]
python = ["37", "38", "39", "310", "311"]

[tool.coverage.run]
branch = true
parallel = true
omit = [
  "mollie_cli/__about__.py",
]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
